# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an auto-generated API Docs.",
        "title": "MyPoints API",
        "termsOfService": "https://swagger.io/terms/",
        "contact": {
            "name": "MyPoints Team Support",
            "email": "zenhanprogram@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/:id/agent": {
            "post": {
                "description": "create agent account by admins.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent"
                ],
                "summary": "admins can create agent account with this api",
                "parameters": [
                    {
                        "description": "body request",
                        "name": "signUp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_yossdev_mypoints-rest-api_src_agents_dto.SignUpReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github.com_yossdev_mypoints-rest-api_src_agents_dto.AccountCreated"
                        }
                    }
                }
            }
        },
        "/admin/login": {
            "post": {
                "description": "check admins by checking given email and password.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "check admins by given email return jwt token if successfully signIn",
                "parameters": [
                    {
                        "description": "body request",
                        "name": "signIn",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_yossdev_mypoints-rest-api_src_admins_dto.SignInReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.Token"
                        }
                    }
                }
            }
        },
        "/admin/profile/:id": {
            "put": {
                "description": "update admin data by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "update admin data",
                "parameters": [
                    {
                        "description": "body request",
                        "name": "updateAccount",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_yossdev_mypoints-rest-api_src_admins_dto.UpdateAccount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github.com_yossdev_mypoints-rest-api_src_admins_dto.AccountUpdated"
                        }
                    }
                }
            }
        },
        "/admin/profile/avatar/:id": {
            "put": {
                "description": "update admin photo profile by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "update admin photo profile",
                "parameters": [
                    {
                        "description": "body request",
                        "name": "updateAvatar",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_yossdev_mypoints-rest-api_src_admins_dto.UpdateAvatar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github.com_yossdev_mypoints-rest-api_src_admins_dto.AccountUpdated"
                        }
                    }
                }
            }
        },
        "/admin/signup": {
            "post": {
                "description": "create admin account.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "admins can create from register page",
                "parameters": [
                    {
                        "description": "body request",
                        "name": "signUp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_yossdev_mypoints-rest-api_src_admins_dto.SignUpReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github.com_yossdev_mypoints-rest-api_src_admins_dto.AccountCreated"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "check agent by checking given email and password.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent"
                ],
                "summary": "check agent by given email return jwt token if successfully signIn",
                "parameters": [
                    {
                        "description": "body request",
                        "name": "signIn",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_yossdev_mypoints-rest-api_src_agents_dto.SignInReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.Token"
                        }
                    }
                }
            }
        },
        "/profile/:id": {
            "get": {
                "description": "Get agent data by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent"
                ],
                "summary": "get agent data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Profile"
                        }
                    }
                }
            },
            "put": {
                "description": "update agent data by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent"
                ],
                "summary": "update agent data",
                "parameters": [
                    {
                        "description": "body request",
                        "name": "updateAccount",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_yossdev_mypoints-rest-api_src_agents_dto.UpdateAccount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github.com_yossdev_mypoints-rest-api_src_agents_dto.AccountUpdated"
                        }
                    }
                }
            }
        },
        "/profile/avatar/:id": {
            "put": {
                "description": "update agent photo profile by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent"
                ],
                "summary": "update agent photo profile",
                "parameters": [
                    {
                        "description": "body request",
                        "name": "updateAvatar",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_yossdev_mypoints-rest-api_src_agents_dto.UpdateAvatar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github.com_yossdev_mypoints-rest-api_src_agents_dto.AccountUpdated"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.Token": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "dto.Profile": {
            "type": "object",
            "properties": {
                "admin_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github.com_yossdev_mypoints-rest-api_src_admins_dto.AccountCreated": {
            "type": "object",
            "properties": {
                "rows_affected": {
                    "type": "integer"
                }
            }
        },
        "github.com_yossdev_mypoints-rest-api_src_admins_dto.AccountUpdated": {
            "type": "object",
            "properties": {
                "rows_affected": {
                    "type": "integer"
                }
            }
        },
        "github.com_yossdev_mypoints-rest-api_src_admins_dto.SignInReq": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github.com_yossdev_mypoints-rest-api_src_admins_dto.SignUpReq": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github.com_yossdev_mypoints-rest-api_src_admins_dto.UpdateAccount": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github.com_yossdev_mypoints-rest-api_src_admins_dto.UpdateAvatar": {
            "type": "object",
            "required": [
                "img"
            ],
            "properties": {
                "img": {
                    "type": "string"
                }
            }
        },
        "github.com_yossdev_mypoints-rest-api_src_agents_dto.AccountCreated": {
            "type": "object",
            "properties": {
                "rows_affected": {
                    "type": "integer"
                }
            }
        },
        "github.com_yossdev_mypoints-rest-api_src_agents_dto.AccountUpdated": {
            "type": "object",
            "properties": {
                "rows_affected": {
                    "type": "integer"
                }
            }
        },
        "github.com_yossdev_mypoints-rest-api_src_agents_dto.SignInReq": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github.com_yossdev_mypoints-rest-api_src_agents_dto.SignUpReq": {
            "type": "object",
            "required": [
                "admin_id",
                "email",
                "name",
                "password"
            ],
            "properties": {
                "admin_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "github.com_yossdev_mypoints-rest-api_src_agents_dto.UpdateAccount": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github.com_yossdev_mypoints-rest-api_src_agents_dto.UpdateAvatar": {
            "type": "object",
            "required": [
                "img"
            ],
            "properties": {
                "img": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}